chore: move description resolver in metadata
